rename index.js to Item.js
