+++ src/components/index.js
@@ -1,2 +1,2 @@
-export {default as Item} from './item';
+export {default as Item} from './item/Item';
 export {default as List} from './list/List';
